Move common code and tests to plugin file.
