+++ sensu/gohandler_test.go
@@ -16,14 +16,14 @@ type handlerValues struct {
 }
 
 var (
-	defaultHandlerConfig = HandlerConfig{
+	defaultHandlerConfig = PluginConfig{
 		Name:     "TestHandler",
 		Short:    "Short Description",
 		Timeout:  10,
 		Keyspace: "sensu.io/plugins/segp/config",
 	}
 
-	defaultOption1 = HandlerConfigOption{
+	defaultOption1 = PluginConfigOption{
 		Argument:  "arg1",
 		Default:   "Default1",
 		Env:       "ENV_1",
@@ -32,7 +32,7 @@ var (
 		Usage:     "First argument",
 	}
 
-	defaultOption2 = HandlerConfigOption{
+	defaultOption2 = PluginConfigOption{
 		Argument:  "arg2",
 		Default:   uint64(33333),
 		Env:       "ENV_2",
@@ -41,7 +41,7 @@ var (
 		Usage:     "Second argument",
 	}
 
-	defaultOption3 = HandlerConfigOption{
+	defaultOption3 = PluginConfigOption{
 		Argument:  "arg3",
 		Default:   false,
 		Env:       "ENV_3",
@@ -73,70 +73,7 @@ func TestNewGoHandler(t *testing.T) {
 	assert.NotNil(t, goHandler.cmdArgs)
 }
 
-func TestSetOptionValue_String(t *testing.T) {
-	finalValue := ""
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "abc")
-	assert.Nil(t, err)
-	assert.Equal(t, "abc", finalValue)
-}
-
-func TestSetOptionValue_EmptyString(t *testing.T) {
-	finalValue := ""
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "")
-	assert.Nil(t, err)
-	assert.Equal(t, "", finalValue)
-}
-
-func TestSetOptionValue_ValidUint64(t *testing.T) {
-	var finalValue uint64
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "123")
-	assert.Nil(t, err)
-	assert.Equal(t, uint64(123), finalValue)
-}
-
-func TestSetOptionValue_InvalidUint64(t *testing.T) {
-	var finalValue uint64
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "abc")
-	assert.NotNil(t, err)
-	assert.Equal(t, uint64(0), finalValue)
-}
-
-func TestSetOptionValue_TrueBool(t *testing.T) {
-	var finalValue bool
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "true")
-	assert.Nil(t, err)
-	assert.Equal(t, true, finalValue)
-}
-
-func TestSetOptionValue_FalseBool(t *testing.T) {
-	finalValue := true
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "false")
-	assert.Nil(t, err)
-	assert.Equal(t, false, finalValue)
-}
-
-func TestSetOptionValue_InvalidBool(t *testing.T) {
-	var finalValue bool
-	option := defaultOption1
-	option.Value = &finalValue
-	err := setOptionValue(&option, "yes")
-	assert.NotNil(t, err)
-	assert.Equal(t, false, finalValue)
-}
-
-func goHandlerExecuteUtil(t *testing.T, handlerConfig *HandlerConfig, eventFile string, cmdLineArgs []string,
+func goHandlerExecuteUtil(t *testing.T, handlerConfig *PluginConfig, eventFile string, cmdLineArgs []string,
 	validationFunction func(*types.Event) error, executeFunction func(*types.Event) error,
 	expectedValue1 interface{}, expectedValue2 interface{}, expectedValue3 interface{}) error {
 	options := getDefaultOptions()
@@ -621,9 +558,9 @@ func clearEnvironment() {
 	_ = os.Unsetenv("ENV_3")
 }
 
-func getDefaultOptions() []*HandlerConfigOption {
+func getDefaultOptions() []*PluginConfigOption {
 	option1 := defaultOption1
 	option2 := defaultOption2
 	option3 := defaultOption3
-	return []*HandlerConfigOption{&option1, &option2, &option3}
+	return []*PluginConfigOption{&option1, &option2, &option3}
 }
